mysql-connector-c++: update to 8.0.25.
MySQL Connectors and other MySQL client tools and applications now
synchronize the first digit of their version number with the
(highest) MySQL server version they support. For example, MySQL
Connector/C++ 8.0.12 would be designed to support all features of
MySQL server version 8 (or lower). This change makes it easy and
intuitive to decide which client version to use for which server
version.