add graphics/fswebcam
fswebcam is a neat and simple webcam app.

It captures images from a V4L1/V4L2 compatible device or file,
averages them to reduce noise and draws a caption using the GD
Graphics Library which also handles compressing the image to PNG
or JPEG. The resulting image is saved to a file or sent to stdio
where it can be piped to something like ncftpput or scp.